+++ logs_sdk/test/opentelemetry/sdk/logs/export/batch_log_record_processor_test.rb
@@ -447,7 +447,7 @@ describe OpenTelemetry::SDK::Logs::Export::BatchLogRecordProcessor do
       exporter = TestExporter.new
       processor = BatchLogRecordProcessor.new(exporter)
 
-      producers = 10.times.map do |i|
+      producers = Array.new(10) do |i|
         Thread.new do
           x = i * 10
           10.times do |j|
@@ -461,7 +461,7 @@ describe OpenTelemetry::SDK::Logs::Export::BatchLogRecordProcessor do
 
       out = exporter.batches.flatten.map(&:body).sort
 
-      expected = 100.times.map { |i| i }
+      expected = Array.new(100) { |i| i }
 
       _(out).must_equal(expected)
     end
@@ -472,6 +472,9 @@ describe OpenTelemetry::SDK::Logs::Export::BatchLogRecordProcessor do
     let(:processor) { BatchLogRecordProcessor.new(exporter) }
 
     it 'reports export failures' do
+      # This test is unreliable on Windows platforms, but works just fine on others
+      skip if Gem.win_platform?
+
       # skip the work method's behavior, we rely on shutdown to get us to the failures
       processor.stub(:work, nil) do
         mock_logger = Minitest::Mock.new