+++ logs_sdk/test/opentelemetry/sdk/logs/log_record_test.rb
@@ -113,7 +113,7 @@ describe OpenTelemetry::SDK::Logs::LogRecord do
         logger.on_emit(attributes: { 'a' => 'a', 'b' => 'b' })
 
         # Look at the captured output to see if the attributes have been truncated
-        assert_match(/attributes={"b"=>"b"}/, captured_stdout.string)
+        assert_match(/attributes={"b".?=>.?"b"}/, captured_stdout.string)
         refute_match(/"a"=>"a"/, captured_stdout.string)
 
         # Return STDOUT to its normal output
@@ -123,14 +123,14 @@ describe OpenTelemetry::SDK::Logs::LogRecord do
       it 'emits an error message if attribute key is invalid' do
         OpenTelemetry::TestHelpers.with_test_logger do |log_stream|
           logger.on_emit(attributes: { a: 'a' })
-          assert_match(/invalid log record attribute key type Symbol/, log_stream.string)
+          assert_match(/Invalid log record attribute key type Symbol/, log_stream.string)
         end
       end
 
       it 'emits an error message if the attribute value is invalid' do
         OpenTelemetry::TestHelpers.with_test_logger do |log_stream|
           logger.on_emit(attributes: { 'a' => Class.new })
-          assert_match(/invalid log record attribute value type Class/, log_stream.string)
+          assert_match(/Invalid log record attribute value type Class/, log_stream.string)
         end
       end