+++ exporter/otlp-logs/CHANGELOG.md
@@ -1,1 +1,13 @@
 # Release History: opentelemetry-exporter-otlp-logs
+
+### v0.2.1 / 2025-07-29
+
+* FIXED: Support hash bodies and attribute values in OTLP logs exporter
+
+### v0.2.0 / 2025-02-25
+
+- ADDED: Support 3.1 Min Version
+
+### v0.1.0 / 2024-12-04
+
+Initial release.