+++ logs_sdk/CHANGELOG.md
@@ -1,1 +1,14 @@
 # Release History: opentelemetry-logs-sdk
+
+### v0.3.0 / 2025-07-29
+
+* ADDED: Update error message for invalid log record attribute
+* ADDED: Set logs exporter default to OTLP
+
+### v0.2.0 / 2025-02-25
+
+- ADDED: Support 3.1 Min Version
+
+### v0.1.0 / 2024-12-04
+
+Initial release.